Update to debhelper compat level 13
dh compat 13 has been supported since Debian 11 (oldoldstable) and
Ubuntu 18.04 LTS (oldoldold-LTS) and compat levels below 9 are
deprecated. This also only affects building the source package, and
trying to install and run a binary distribution of crawl on really
old versions of Debian/Ubuntu is likely to not work anyways.